fix(hooks): decouple vault and plugin path resolution plugin-wide
Extract vault resolution from wiki-hooks.py into shared scripts/_vault.py
module. All scripts now use the 5-step resolution chain ($CLAUDE_OBSIDIAN_VAULT,
CLAUDE.md in project dir, CLAUDE.md in cwd, walk-up, user config) instead of
hardcoding Path(__file__).parent.parent.

Plugin root resolves from $CLAUDE_PLUGIN_ROOT (set by Claude Code) with
fallback to script parent for standalone use.

- New: scripts/_vault.py (shared resolve_vault + resolve_vault_or_die)
- New: skills/wiki/references/vault-resolution.md (LLM-facing reference)
- Refactored: wiki-hooks.py imports from _vault (-173 lines)
- Refactored: tiling-check.py, boundary-score.py use _vault
- Updated: 8 SKILL.md files with vault resolution preamble
- Fixed: all bash snippets use $CLAUDE_PLUGIN_ROOT and $VAULT_ROOT
