+++ skills/canvas/SKILL.md
@@ -6,6 +6,8 @@ allowed-tools: Read Write Edit Glob Grep
 
 # canvas: Visual Reference Layer
 
+> **Vault resolution**: The vault and plugin are separate directories. Before using any paths, read `skills/wiki/references/vault-resolution.md` to resolve `VAULT_ROOT` (where `wiki/` lives). All `wiki/` paths below are relative to `VAULT_ROOT`.
+
 The three knowledge capture layers:
 - `/save` → text synthesis (wiki/questions/, wiki/concepts/)
 - `/autoresearch` → structured knowledge (wiki/sources/, wiki/concepts/)
