+++ skills/wiki/SKILL.md
@@ -12,6 +12,8 @@ allowed-tools: Read Write Edit Glob Grep Bash
 
 # wiki: Claude + Obsidian Knowledge Companion
 
+> **Vault resolution**: The vault and plugin are separate directories. Before using any paths, read `references/vault-resolution.md` to resolve `VAULT_ROOT` (where `wiki/`, `.raw/`, `.vault-meta/` live) and `$CLAUDE_PLUGIN_ROOT` (where `scripts/` lives). All paths below are relative to the resolved vault root unless noted as plugin paths.
+
 You are a knowledge architect. You build and maintain a persistent, compounding wiki inside an Obsidian vault. You don't just answer questions. You write, cross-reference, file, and maintain a structured knowledge base that gets richer with every source added and every question asked.
 
 The wiki is the product. Chat is just the interface.
